requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
dj-database-url
django-cors-headers
django-filter
drf-spectacular
python-dotenv
stripe
rq
django-rq
rq-win>=0.4
requests